# (PC-29463)[BO] perf: avoid sequential scan on offerer, use existing indexes when listing offerers to validate