## But de la pull request

Ticket Jira : https://passculture.atlassian.net/browse/PC-29463

According to pg_stat_user_tables, index ix_offerer_trgm_unaccent_city was never used.
Actually, a sequential scan was still made on offerers to validate.

Before:

```
EXPLAIN ANALYZE SELECT
    offerer.address AS offerer_address, offerer.street AS offerer_street, offerer.id AS offerer_id, offerer."isActive" AS "offerer_isActive", offerer."validationStatus" AS "offerer_validationStatus", offerer."postalCode" AS "offerer_postalCode", offerer.city AS offerer_city, offerer."dateCreated" AS "offerer_dateCreated", offerer.name AS offerer_name, offerer.siren AS offerer_siren, offerer."dateValidated" AS "offerer_dateValidated", offerer."allowedOnAdage" AS "offerer_allowedOnAdage", "user".id AS creator_id, "user".email AS creator_email
FROM offerer
LEFT OUTER JOIN user_offerer ON user_offerer.id = (
    SELECT user_offerer.id
    FROM user_offerer
    WHERE user_offerer."offererId" = offerer.id
    ORDER BY user_offerer.id ASC
    LIMIT 1
)
LEFT OUTER JOIN "user" ON "user".id = user_offerer."userId" WHERE (
    immutable_unaccent(offerer.name) ILIKE '%ma ville%'
    OR immutable_unaccent(offerer.city) ILIKE '%ma ville%'
    OR immutable_unaccent("user"."firstName" || ' ' || "user"."lastName") ILIKE '%ma ville%'
)
AND offerer."validationStatus" IN ('NEW', 'PENDING', 'VALIDATED', 'REJECTED') ;

+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------->
| QUERY PLAN                                                                                                                                                                                                                                                                 >
|---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------->
| Hash Left Join  (cost=156812.19..383584.31 rows=12 width=166) (actual time=220.323..1080.798 rows=190 loops=1)                                                                                                                                                             >
|   Hash Cond: ((SubPlan 1) = user_offerer.id)                                                                                                                                                                                                                               >
|   Filter: ((immutable_unaccent((offerer.name)::text) ~~* '%ma ville%'::text) OR (immutable_unaccent((offerer.city)::text) ~~* '%ma ville%'::text) OR (immutable_unaccent(((("user"."firstName")::text || ' '::text) || ("user"."lastName")::text)) ~~* '%ma ville%'::text))>
|   Rows Removed by Filter: 42205                                                                                                                                                                                                                                            >
|   ->  Seq Scan on offerer  (cost=0.00..2237.12 rows=45008 width=130) (actual time=0.018..29.096 rows=42395 loops=1)                                                                                                                                                        >
|         Filter: ("validationStatus" = ANY ('{NEW,PENDING,VALIDATED,REJECTED}'::validationstatus[]))                                                                                                                                                                        >
|   ->  Hash  (cost=155626.88..155626.88 rows=50985 width=57) (actual time=168.634..168.751 rows=49529 loops=1)                                                                                                                                                              >
|         Buckets: 65536  Batches: 2  Memory Usage: 2720kB                                                                                                                                                                                                                   >
|         ->  Gather Merge  (cost=1000.75..155626.88 rows=50985 width=57) (actual time=9.480..135.283 rows=49529 loops=1)                                                                                                                                                    >
|               Workers Planned: 2                                                                                                                                                                                                                                           >
|               Workers Launched: 2                                                                                                                                                                                                                                          >
|               ->  Nested Loop Left Join  (cost=0.72..148741.92 rows=21244 width=57) (actual time=0.075..116.264 rows=16510 loops=3)                                                                                                                                        >
|                     ->  Parallel Index Only Scan using user_offerer_pkey on user_offerer  (cost=0.29..2820.52 rows=21244 width=16) (actual time=0.041..11.434 rows=16510 loops=3)                                                                                          >
|                           Heap Fetches: 49529                                                                                                                                                                                                                              >
|                     ->  Index Scan using user_pkey on "user"  (cost=0.43..6.87 rows=1 width=49) (actual time=0.006..0.006 rows=1 loops=49529)                                                                                                                              >
|                           Index Cond: (id = user_offerer."userId")                                                                                                                                                                                                         >
|   SubPlan 1                                                                                                                                                                                                                                                                >
|     ->  Limit  (cost=8.32..8.32 rows=1 width=8) (actual time=0.004..0.004 rows=1 loops=81929)                                                                                                                                                                              >
|           ->  Sort  (cost=8.32..8.32 rows=1 width=8) (actual time=0.003..0.003 rows=1 loops=81929)                                                                                                                                                                         >
|                 Sort Key: user_offerer_1.id                                                                                                                                                                                                                                >
|                 Sort Method: quicksort  Memory: 25kB                                                                                                                                                                                                                       >
|                 ->  Index Scan using "ix_user_offerer_offererId" on user_offerer user_offerer_1  (cost=0.29..8.31 rows=1 width=8) (actual time=0.002..0.002 rows=1 loops=81929)                                                                                            >
|                       Index Cond: ("offererId" = offerer.id)                                                                                                                                                                                                               >
| Planning Time: 5.359 ms                                                                                                                                                                                                                                                    >
| Execution Time: 1081.172 ms                                                                                                                                                                                                                                                >
+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------->
```

```
After:

EXPLAIN ANALYZE SELECT
    offerer.address AS offerer_address, offerer.street AS offerer_street, offerer.id AS offerer_id, offerer."isActive" AS "offerer_isActive", offerer."validationStatus" AS "offerer_validationStatus", offerer."postalCode" AS "offerer_postalCode", offerer.city AS offerer_city, offerer."dateCreated" AS "offerer_dateCreated", offerer.name AS offerer_name, offerer.siren AS offerer_siren, offerer."dateValidated" AS "offerer_dateValidated", offerer."allowedOnAdage" AS "offerer_allowedOnAdage", "user".id AS creator_id, "user".email AS creator_email
FROM offerer
LEFT OUTER JOIN user_offerer ON user_offerer.id = (
    SELECT user_offerer.id
    FROM user_offerer
    WHERE user_offerer."offererId" = offerer.id
    ORDER BY user_offerer.id ASC
    LIMIT 1
)
LEFT OUTER JOIN "user" ON "user".id = user_offerer."userId" WHERE (
    immutable_unaccent(offerer.name) ILIKE '%ma ville%'
    OR immutable_unaccent(offerer.city) ILIKE '%ma ville%'
    --OR immutable_unaccent("user"."firstName" || ' ' || "user"."lastName") ILIKE '%ma ville%'
)
AND offerer."validationStatus" IN ('NEW', 'PENDING', 'VALIDATED', 'REJECTED') UNION
SELECT
    offerer.address AS offerer_address, offerer.street AS offerer_street, offerer.id AS offerer_id, offerer."isActive" AS "offerer_isActive", offerer."validationStatus" AS "offerer_validationStatus", offerer."postalCode" AS "offerer_postalCode", offerer.city AS offerer_city, offerer."dateCreated" AS "offerer_dateCreated", offerer.name AS offerer_name, offerer.siren AS offerer_siren, offerer."dateValidated" AS "offerer_dateValidated", offerer."allowedOnAdage" AS "offerer_allowedOnAdage", "user".id AS creator_id, "user".email AS creator_email
FROM offerer
LEFT OUTER JOIN user_offerer ON user_offerer.id = (
    SELECT user_offerer.id
    FROM user_offerer
    WHERE user_offerer."offererId" = offerer.id
    ORDER BY user_offerer.id ASC
    LIMIT 1
)
LEFT OUTER JOIN "user" ON "user".id = user_offerer."userId" WHERE (
    --immutable_unaccent(offerer.name) ILIKE '%ma ville%'
    --OR immutable_unaccent(offerer.city) ILIKE '%ma ville%'
    --OR
    immutable_unaccent("user"."firstName" || ' ' || "user"."lastName") ILIKE '%ma ville%'
)
AND offerer."validationStatus" IN ('NEW', 'PENDING', 'VALIDATED', 'REJECTED') ;

+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------->
| QUERY PLAN                                                                                                                                                                                                                                                                 >
|---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------->
| HashAggregate  (cost=103145.63..103145.75 rows=12 width=1226) (actual time=262.991..263.266 rows=190 loops=1)                                                                                                                                                              >
|   Group Key: offerer.address, offerer.street, offerer.id, offerer."isActive", offerer."validationStatus", offerer."postalCode", offerer.city, offerer."dateCreated", offerer.name, offerer.siren, offerer."dateValidated", offerer."allowedOnAdage", "user".id, "user".emai>
|   ->  Append  (cost=617.10..103145.21 rows=12 width=1226) (actual time=3.112..262.462 rows=190 loops=1)                                                                                                                                                                    >
|         ->  Nested Loop Left Join  (cost=617.10..802.86 rows=7 width=166) (actual time=3.111..6.939 rows=190 loops=1)                                                                                                                                                      >
|               ->  Nested Loop Left Join  (cost=616.67..754.77 rows=7 width=138) (actual time=3.092..6.409 rows=190 loops=1)                                                                                                                                                >
|                     ->  Bitmap Heap Scan on offerer  (cost=608.06..638.29 rows=7 width=130) (actual time=3.041..5.064 rows=190 loops=1)                                                                                                                                    >
|                           Recheck Cond: ((immutable_unaccent((name)::text) ~~* '%ma ville%'::text) OR (immutable_unaccent((city)::text) ~~* '%ma ville%'::text))                                                                                                           >
|                           Rows Removed by Index Recheck: 2                                                                                                                                                                                                                 >
|                           Filter: ("validationStatus" = ANY ('{NEW,PENDING,VALIDATED,REJECTED}'::validationstatus[]))                                                                                                                                                      >
|                           Heap Blocks: exact=19                                                                                                                                                                                                                            >
|                           ->  BitmapOr  (cost=608.06..608.06 rows=7 width=0) (actual time=2.894..2.896 rows=0 loops=1)                                                                                                                                                     >
|                                 ->  Bitmap Index Scan on ix_offerer_trgm_unaccent_name  (cost=0.00..376.03 rows=4 width=0) (actual time=1.785..1.786 rows=4 loops=1)                                                                                                       >
|                                       Index Cond: (immutable_unaccent((name)::text) ~~* '%ma ville%'::text)                                                                                                                                                                >
|                                 ->  Bitmap Index Scan on ix_offerer_trgm_unaccent_city  (cost=0.00..232.02 rows=3 width=0) (actual time=1.107..1.107 rows=201 loops=1)                                                                                                     >
|                                       Index Cond: (immutable_unaccent((city)::text) ~~* '%ma ville%'::text)                                                                                                                                                                >
|                     ->  Index Only Scan using user_offerer_pkey on user_offerer  (cost=8.61..16.63 rows=1 width=16) (actual time=0.002..0.002 rows=1 loops=190)                                                                                                            >
|                           Index Cond: (id = (SubPlan 1))                                                                                                                                                                                                                   >
|                           Heap Fetches: 190                                                                                                                                                                                                                                >
|                           SubPlan 1                                                                                                                                                                                                                                        >
|                             ->  Limit  (cost=8.32..8.32 rows=1 width=8) (actual time=0.004..0.004 rows=1 loops=190)                                                                                                                                                        >
|                                   ->  Sort  (cost=8.32..8.32 rows=1 width=8) (actual time=0.003..0.003 rows=1 loops=190)                                                                                                                                                   >
|                                         Sort Key: user_offerer_2.id                                                                                                                                                                                                        >
|                                         Sort Method: quicksort  Memory: 25kB                                                                                                                                                                                               >
|                                         ->  Index Scan using "ix_user_offerer_offererId" on user_offerer user_offerer_2  (cost=0.29..8.31 rows=1 width=8) (actual time=0.002..0.002 rows=1 loops=190)                                                                      >
|                                               Index Cond: ("offererId" = offerer.id)                                                                                                                                                                                       >
|               ->  Index Scan using user_pkey on "user"  (cost=0.43..6.87 rows=1 width=36) (actual time=0.002..0.002 rows=1 loops=190)                                                                                                                                      >
|                     Index Cond: (id = user_offerer."userId")                                                                                                                                                                                                               >
…